split js files
