+++ webpack.config.js
@@ -6,7 +6,7 @@ module.exports = {
   context: path.join(__dirname, '/src/model'),
   entry: { 
     calculator: './index.js',
-    test: "./index.test.js"
+    //test: "./index.test.js"
   },
 
   output: {
